pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "gspread>=6.1.4",
    "oauth2client>=4.1.3",
    "pandas>=2.2.3",
    "python-telegram-bot>=20.0",
]
